KL-145/rename: split into new file
- useQuery, useMutation 다른 파일로 각각 분리
