+++ src/hooks/useKyQuery.js
@@ -1,20 +1,5 @@
-import ky from 'ky'
-import { useQueryClient, useQuery, useMutation } from '@tanstack/react-query'
-
-const method = {
-  GET: 'get',
-  POST: 'post',
-  PUT: 'put',
-  DELETE: 'delete',
-}
-
-const kyInstance = ky.create({
-  prefixUrl: 'http://localhost:8080/v1/',
-  timeout: 5000,
-  headers: {
-    'content-type': 'application/json',
-  },
-})
+import { useQuery } from '@tanstack/react-query'
+import { kyInstance } from './kyInstance'
 
 /**
  * Ky기반 useQeury 호출 훅
@@ -34,26 +19,4 @@ const useKyQuery = (uri, options = null) =>
     ...options,
   })
 
-/**
- * Ky기반 useMutation 호출 훅
- *
- * @param {string} httpMethod http 메서드 (post, put, patch, delete)
- * @param {string} uri 요청할 uri
- * @param {string} body post, put, patch, delete
- * @param {object} options 요청시 넣을 옵션 (default: null)
- * @return useQuery의 return과 동일
- */
-const useKyMutation = (httpMethod, uri, options = null) => {
-  const queryClient = useQueryClient()
-
-  return useMutation({
-    mutationFn: (body) => kyInstance[httpMethod](uri, { body }).json(),
-    onSuccess: (data) => {
-      queryClient.invalidateQueries([uri, data.productId])
-    },
-    retry: false,
-    ...options,
-  })
-}
-
-export { method, kyInstance, useKyQuery, useKyMutation }
+export default useKyQuery
